+++ src/components/Calculator/Calculator.js
@@ -1,0 +1,35 @@
+import React, { useState } from 'react';
+import Screen from './Screen';
+import BoxButton from './BoxButton';
+import Button from './Button';
+import calculate from '../../logic/calculate';
+
+const btnValues = [
+  ['AC', '+/-', '%', '÷'],
+  [7, 8, 9, '×'],
+  [4, 5, 6, '-'],
+  [1, 2, 3, '+'],
+  [0, '.', '='],
+];
+
+const Calculator = () => {
+  const [calc, setCalc] = useState({});
+  const { next, total } = calc;
+
+  return (
+    <div className="calc-container">
+      <Screen next={next} total={total} />
+      <BoxButton>
+        {btnValues.flat().map((btn) => (
+          <Button
+            value={btn.toString()}
+            key={btn}
+            onClick={() => setCalc(calculate(calc, btn.toString()))}
+          />
+        ))}
+      </BoxButton>
+    </div>
+  );
+};
+
+export default Calculator;
