+++ src/components/Quote/Quote.js
